+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - [`compete.toml`] Changed the behavior of `exclude`.
+- [rename] Renamed `migrate packages` command to `migrate cargo-atcoder`.
 
 ### Added
 
@@ -12,7 +13,7 @@
 
 ### Fixed
 
-- [`migrate` command] Now `cargo compete migrate` also generates a template package as `cargo compete init` does.
+- [`migrate cargo-atcoder`] Now `cargo compete migrate` also generates a template package as `cargo compete init` does.
 
 ## [0.3.2] - 2020-08-08Z